moved: ProjectQueryHelper
